modifing project

+++ imageobject.cpp
@@ -1,7 +1,7 @@
 #include "imageobject.h"
 #include <QImage>
 
-ImageObject::ImageObject(QImage current_image, int current_size_factor, int current_rotation_angle)
+ImageObject::ImageObject(QImage current_image, double current_size_factor, int current_rotation_angle)
 {
     this->image = current_image;
     this->size_factor = current_size_factor;
